Consolidated GLBMuonAnalyzer into one file
The module was never actually registered with the plugin manager
as SealModule.cc was not compiled.
This was changed to remove an unneeded dependency on
FWCore/Services/src/SiteLocalConfigService.h.
